+++ src/wind_simulation.py
@@ -3,13 +3,28 @@ from generation_algorithm import Node
 from generation_algorithm import print_table
 class WindSimulator:
     def __init__(self, table):
-        #initializes the old table and an empty new table
+        """initializes the old table and an empty new table
+
+        Args:
+            table: This is the previous frame of the cloud animation.
+        """
         self.old_table = table
         self.new_table = []
         for _ in range(0, len(self.old_table)):
             self.new_table.append([None]*len(self.old_table[0]))
 
     def find_new_coords(self, col_i, row_i, row):
+        """This function find the new coordinates of a node
+        based on the give previous coordinates
+
+        Args:
+            col_i : The x coordinate of node in the previous frame.
+            row_i : The y coordinate of node in the previous frame.
+            row : The current row in the previous frame.
+
+        Returns:
+            _type_: _description_
+        """
         y_move=0
         x_move=0
         x_dif = col_i
@@ -22,6 +37,7 @@ class WindSimulator:
 
         left_side = col_i <= len(row)//2
         up_side = row_i <= len(self.old_table)//2
+
         if not(len(self.old_table)%2!=0 and
         abs(len(self.old_table)-1 - row_i) == row_i and
         x_dif>=y_dif):
@@ -40,6 +56,11 @@ class WindSimulator:
         return [row_i + y_move, col_i + x_move]
 
     def simulate(self):
+        """This function uses find_new_coords for every node to create the new frame.
+
+        Returns:
+            new_table: This is the newly generated frame of the animation.
+        """
         for row_i, row in enumerate(self.old_table):
             for col_i, col in enumerate(row):
                 if not col:
@@ -49,15 +70,20 @@ class WindSimulator:
                 col.coords = new_coords
                 col.neighbors.clear()
                 self.new_table[new_coords[0]][new_coords[1]] = col
+        for row_i, row in enumerate(self.new_table):
+            for col_i, col in enumerate(row):
+                if not col:
+                    continue
                 col.find_neighbors()
+
         return self.new_table
 
 
 if __name__ == "__main__": # pragma: no cover
     table_1 = []
     frames=[]
-    N = 13
-    M = 20
+    N = 20
+    M = 40
     for _ in range(0, N):
         table_1.append([None]*M)
     new_node = Node([(len(table_1)-1)//2,(len(table_1[0])-1)//2], table_1)